requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
docopt==0.6.2
idna==3.3
pipreqs==0.4.11
pycparser==2.20
pymongo==3.12.1
PyYAML==6.0
requests==2.26.0
urllib3==1.26.7
xattr==0.9.7
yarg==0.1.9
